Update contact info and restructure team members
Updated contact details and address in both English and Chinese locales, and changed the email in ContactPage. Moved '余浩铭' and '李一楠' from developers/designers to contributors in team constants. Renamed a contributor image file. Also, moved the NEC application card to the contact form section for improved layout. Updated a project title in ProjectsSection.